requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.2
pycryptodome>=3.19.0
reportlab>=4.0.8
Pillow>=10.2.0
streamlit>=1.32.0
PyMuPDF>=1.23.8
pandas>=2.2.0
numpy>=1.26.0
pytest>=8.0.0 